refactor: 전략이 변경될 사항이 없어 strategy 패턴 제거
