--- src/main/java/racewinner/domain/strategy/MoveStrategy.java
@@ -1,5 +1,0 @@
-package racewinner.domain.strategy;
-
-public interface MoveStrategy {
-    int number();
-}
